+++ tests/test_csv_row.cpp
@@ -16,7 +16,7 @@ TEST_CASE("CSVRow Test", "[test_csv_row]") {
         "Col3"
         "Col4";
 
-    std::vector<size_t> splits = { 4, 8, 12 };
+    std::vector<unsigned short> splits = { 4, 8, 12 };
 
     CSVRow row(
         std::move(str),
@@ -77,7 +77,7 @@ TEST_CASE("CSVField operator==", "[test_csv_field_equal]") {
         "3"
         "3.14";
 
-    std::vector<size_t> splits = { 1, 2, 3 };
+    std::vector<unsigned short> splits = { 1, 2, 3 };
     CSVRow row(std::move(str), std::move(splits), col_names);
 
     REQUIRE(row["A"] == 1);
